feat(util-prop-types): add responsiveProps
