+++ packages/prop-types/package.json
@@ -19,5 +19,8 @@
   "bugs": {
     "url": "https://github.com/telusdigital/tds/issues"
   },
-  "homepage": "http://tds.telus.com"
+  "homepage": "http://tds.telus.com",
+  "dependencies": {
+    "prop-types": "^15.6.2"
+  }
 }
